refactor(condo): DOMA-4226 moved `sbbolSecretStorage` instance to `singletons.js` module
