+++ apps/condo/domains/organization/integrations/sbbol/sync/index.js
@@ -4,8 +4,8 @@ const { RUSSIA_COUNTRY } = require('@condo/domains/common/constants/countries')
 const { normalizePhone } = require('@condo/domains/common/utils/phone')
 const { normalizeEmail } = require('@condo/domains/common/utils/mail')
 const { getOrganizationEmployee } = require('@condo/domains/organization/utils/serverSchema/OrganizationEmployee')
-
-const { SBBOL_IMPORT_NAME, sbbolSecretStorage } = require('../common')
+const { SBBOL_IMPORT_NAME } = require('../common')
+const { sbbolSecretStorage } = require('../singletons')
 const { dvSenderFields } = require('../constants')
 const { syncUser } = require('./syncUser')
 const { syncOrganization } = require('./syncOrganization')
