+++ apps/condo/domains/organization/integrations/sbbol/sync/syncTokens.js
@@ -1,4 +1,4 @@
-const { sbbolSecretStorage } = require('../common')
+const { sbbolSecretStorage } = require('../singletons')
 
 /**
  *
