+++ apps/condo/domains/organization/integrations/sbbol/utils/changeClientSecret.js
@@ -1,7 +1,7 @@
 const querystring = require('querystring')
 const dayjs = require('dayjs')
 const conf = require('@condo/config')
-const { sbbolSecretStorage } = require('../common')
+const { sbbolSecretStorage } = require('../singletons')
 const { SbbolRequestApi } = require('../SbbolRequestApi')
 const { getOrganizationAccessToken } = require('./getOrganizationAccessToken')
