+++ apps/condo/domains/organization/integrations/sbbol/utils/getOrganizationAccessToken.js
@@ -1,4 +1,4 @@
-const { sbbolSecretStorage } = require('../common')
+const { sbbolSecretStorage } = require('../singletons')
 const { SbbolOauth2Api } = require('../oauth2')
 
 /**
